Show folder icons for the locations
